Fix Issue With RND Instruction
Previously we were taking the modulo instead of the logical AND of the
immediate value given. These two ways of truncating a random number are not
equivalent and using the modulus was producing strange issues with several
games including Tetris. By changing the modulo to logical AND those issues are
resolved.